Updated the Gherkin Parser to 2.12.1 from 2.6.5.
The primary motivation is that there are some Scenario Outline issues that bothers me and I wanted to make sure that the Parser was not to blame.

Updating the parser wasn't exactly trivial since the storage location and format of the language information has changed considerably since the old version 2.6.5. It is now a json-file (i18n.json) located in a gherkin.jar-resource in the Gherkin.dll.

To ease future updates of the Gherkin Parser I have done the following:

  - Added Gherkin as a Nuget package to ImportGherkinParser.csproj
  - Changed Program.cs in ImportGherkinParser.csproj to reflect the new position of the language information
  - Added Documentation/How to update the Gherkin Parser.txt to explain how to update the parser
  - Added Sample Output from ImportGherkinParser.txt to show what kind of output comes from ImportGherkinParser.csproj
  - Updated CoreFiles.wxs to reflect that various IKVM.* files now are a Nuget package instead of in lib/gherkin
  - Removed various IKVM* files from lib/gherkin
  - Updated ListenerExtender.cs since the Listener interface from Gherkin had changed (not by much, though)