#416 solved
I made a change in which we will first try to load the plugin assembly
as the name states in the config file. If this doesn't succeed a
FileNotFound exception is thrown. If this is the case, in order to
maintain retrocompatibility, the code will try to load the assembly with
the suffix. If even that doesn't succeed, a SpecFlow exception will be
thrown.
It is not the most elegant code, but it is a quick fix for an anoing
bug.