Initial implementation of folder based generator
Since .NET Core uses directories as a container for projects rather than a
csproj file the generator needed some fixing. It now supports the basics
for generating feature files in a folder.

You can use specflow.json to configure some settings. If you don't have it
in the directory the tool will assume that you want to use the defaults.