Explicitly catch errors in the server startup inside tests (#1124)
Adding this code lead to interesting obsevation that all 3 tests is now failed with client not able to connect to server. But for example when  run SingleCall + Shutdown tests where in Shutdown test I disable starting server thread and verification of server creation, SingleCall fails and Shutdown pass. Looks like this is indication of passing tests here reuse server instance from previous tests. I think this could be fixed, but adding property IsRunning which will verify that server is started, and use that property insted of using Thread.Sleep(1000) in the code.