Change contract for registering default value comparers.
Replace RegisterDefault with ReplaceDefault to more clearly indicate
that there can only be one default, and to give the desired behaviour in
one move, without requiring a separate step to remove the original
default.

Remove RegisterDefault from Service: as this has not yet been released,
there is no need to add a method that will be obsolete.