Fix XUnit output (SpecFlow trace information were not displayed in Test Explorer) (#1450)
When tests are executed with xUnit (e.g. with VS Test Explorer), the standard SpecFlow trace information (running steps, etc.) is not included in the output.

The issue is caused by the registration of the XUnitTraceListener, that was registered to the Scenario container, although SpecFlow reads it out from the Test Thread Container. The plugin is changed now to register it to the right container and inside the method it checks if the `ITestOutputHelper` is available and if yes, it forwards the message there. 

Tested manually.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected).

## Checklist:

- [ ] I've added tests for my code.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added an entry to the changelog