Preparation to fix #1348: Expose TestRunnerManager.OnTestRunEnd to be able to call it from plugins (#1458)
In order to fix #1348 in a way as @SabotageAndi suggested, we need an easy way to ensure that the AfterTestRun hooks fired even before AppDomain unload. 

This PR exposes an `OnTestRunEnd` method on the static API of `TestRunnerManager` to support this. The `OnTestRunEnd` method is safe to call multiple times (or call it before domain unload), it will fire the events only once. 

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected).

## Checklist:

- [x] I've added tests for my code.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added an entry to the changelog