Update rules for step definitions method
The Step Definition Methods Rules section states that "method can not have a return type or have a `Task` as return type". I tried to change my method to return `int` instead of `void` and it works without errors. So I guess if it's not a rule which would cause an error than it should be rather recommendation than rule.